generate a SQLite Database
So rather than try to merge many disparate flat file sources, lets
just pretend that the 1970s actually happened and use a SQL engine.

This commit adds a script to generate a SQLite database from the
job postings. It leverages quite a few more dependencies on CPAN
but all in all I think it's worth it in the long run for making this
project a bit more developer friendly.

Additionally we commit the first version of `Perl_Companies.db` the actual
companies database.

Next steps are to generate the various file formats FROM `Perl_Companies.db`
and to look at ways to scrub the data in `Perl_Companies.db`.