import from YAML and reverse the flow
This took a little re-jiggering (hence the big-ass commit). Basically
I didn't want to lose any information that was in the existing YAML file
so I wrote a script to import it into the SQLite DB.

I then re-wrote the scripts to generate the YAML file as well as the
markdown and CSV files from the SQLite DB.

This should complete the *code* portion of this re-write.